simplify AutomationTask class hiearchy; unify task_type (SubTaskType) for all all Task classes; update AutomationTaskInterface;
