+++ graphql_api/schema/custom/rupture_generation_task.py
@@ -20,12 +20,8 @@ from graphql_api.config import CW_METRICS_RESOLUTION, STACK_NAME
 from graphql_api.data import get_data_manager
 from graphql_api.schema.thing import Thing
 
-from .automation_task import AutomationTask
-from .automation_task_base import (  # AutomationTask,
-    AutomationTaskInput,
-    AutomationTaskInterface,
-    AutomationTaskUpdateInput,
-)
+from .automation_task import AutomationTask, AutomationTaskInput, AutomationTaskInterface, AutomationTaskUpdateInput
+from .common import TaskSubType
 
 db_metrics = ServerlessMetricWriter(
     lambda_name=STACK_NAME, metric_name="MethodDuration", resolution=CW_METRICS_RESOLUTION
@@ -39,6 +35,11 @@ class RuptureGenerationTask(AutomationTask):
 
     class Meta:
         interfaces = (relay.Node, Thing, AutomationTaskInterface)
+
+    def resolve_task_type(root, info, **args):
+        if task_type := root.task_type:
+            return task_type
+        return TaskSubType.RUPTURE_SET
 
     @staticmethod
     def from_json(jsondata):
@@ -56,14 +57,6 @@ class RuptureGenerationTaskConnection(relay.Connection):
     @staticmethod
     def resolve_total_count(root, info, *args, **kwargs):
         return len(root.edges)
-
-
-# def json_ready(input):
-#     json_ready_input = copy.copy(input)
-#     for fld in ['result', 'state']:
-#         if json_ready_input.get(fld):
-#             json_ready_input[fld] = json_ready_input[fld].value
-#     return json_ready_input
 
 
 class CreateRuptureGenerationTask(graphene.Mutation):
